Updated the way that parsers were being mixed into the warehouse.
The new way of mixing in a parser fixes the bug which occured when overwriting an existing parser.  The model must be saved first to get this fix which will be changed in the next fix.  Parsers are now set anytime the model is read from cash or the database.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# Committer: Mathew Wilson <daicoden@copypastel.com>
#
# On branch refactoring_and_testing_warehouses
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   app/models/warehouse.rb
#