Finished getting the warehouse parser management to work as expected.
Last commit was accidental, Sorry.  Now when you set warehouse.parser to something the object will be using the proper parser, even if it has loaded a previous one.  We still need to figure out a way to unload the previous parser to pass specs which are to come.

The code is allot more concise now and the ParserManager might not need to be its own module now.  Thoughts welcome!