Added factories to specs.  A factory can take :valid, :saved, :new, and a block.
Thinking about adding some default spec generation.  Inside a spec you would write it_should_have_property(:name) and a spec would be auto-generated to test that name was there.  I would also add it_should_have_property(:name, :unique => true).  This is where the real benefit comes in because it seems like I am writing the same cases for each unique property to ensure it stays unique.

It is also important to note these are not auto-generated from the model so if in the future something changes it will require the editing of the spec to make it work.  Let me know what you think.