Warehouse specs back underway, found some holes in the design while making specs and need to talk with Ecin before proceeding.
The main problem's are: we need a different name for Detail, still don't know what it is. Why are carts and orders different?  They should be viewed as uncompleted orders.  Lastly, why does Warehouse have the class method price?  This is only valuable if a warehouse can look up a manufacture ID or something which involves the item archetypes and we are not there yet, so as a good specer this should be left out!